Add save and load game functions
Save game will dump the class variable, stats, and name into a file, then load game will read from the same file to apply those to the in-game stats, thus allowing a person to save their character. Later, there will be other variables to track game progress (when there actually is a game to play) and thus we will have an effective system of saving and loading the game. For now it just copies user variables.