Add player class
The player is an object, and thus, we make him/her one. All of the functions involving the player (level up, choose class, save/load game, and display) have been moved to the class as well. The code will be much more clean and we can remove the excessive functions from main. The functions have also been cleaned up in the player class.