Documentation and formatting
After the big code push yesterday, I realized it was lacking a LOT of explanations/documentation. So this is the first changes out of a few to catch up and be prepared in case there's any new contributors. Reminder for the next few commits:Finish documentation for player/monster in their headers and add some 'helpful' comments in the actual function code (like battle's code).