GUI Update: gtkmm 3 and more
This is a huge update. Let's break it down:

I've moved from a console implementation to a GUI implementation with gtkmm. gtk+ is cross platform, flexible, and very capable. gtk+ is a C library, so we use gtkmm which provides us a C++ interface.
So far the game starts up, has a main menu and a character creation screen. Upon creating your character it will delete the character creation interface to stage for the next one, but there is none.

Most of the code still needs to be transferred over, as in it is currently still reliant on cout and cin for a user interface. These files that are currently unused have been moved to the unused folder.

I have decided to go with the GPLv3 license for this project due to the fact that I want it to remain open sourced, so you'll see that header in every file now.

Please keep in mind that documentation is incomplete and implementation is incomplete. It DOES compile but that's about it. There should be a lot more text here but this commit has been long overdue anyway, let's just push it out.