Change naming and general structure for Frame/Window objects
First thing's first: naming needs to be consistent. I'm tired of underlines, so
we will now rely on capital letters to indicate word differeces.

Most of the changes here have to do with consistency. Naming and structure are
the two big ones. mainWindow and our frames will now use a Pack Main Unpack
structure going forward. This will make it easier to understand what is
happening and what goes where. All things organization/packing go in the pack
functions, everything else like naming labels and binding signals will be
located in the main (along with the pack call initially). If necessary, there
will be an unpack to perform any necessary calls to unpack the frame from the
window or anything else.

Many of the variables (if not all) in mainWindow have been renamed so it's now
easier to take one look at them and have a general understanding of where they
belong/what they do.

Work on the main game window has begun, so it'll transition to that screen. The
next change should be building upon this so we can actually play the damn game.