Update names and code style
I have decided C++ code looks much better in snake_case. So I've put the code
to a (relatively) consistent snake_case naming convention. I have also added a
coding style md file... which while this project is hardly followed, it is
important to document these things. It is fairly incomplete as I am going to
search for a new style to format my code with. This is as far as the project
goes with naming for now, the next round of updates will hopefully be focused
on implementation of the game.