Fix red box redraw bug
Changed the default color pairing (1) to white and black. This seemed to be the
primary factor in fixing the issue. Also set the main window back to white as
well to ensure colors are in order. To allivate text_log being red, I changed
its color to 1 to keep it's white color as well.