Add pugixml library script
We're going to go with pugixml because it's fast and easy to use. It's covered
under the MIT license, so therefore we have no issues using it. Also, it is
very well written from what I have seen.

The script will grab the pugixml source, extract the relevant header files,
compile the implementation file, and pack it as a library into the lib folder.
You'll notice we are not tracking the lib folder from our .gitignore. This is
so the github continues to track only files I have written. Afterwards it
removes all "extra" files.

I will, of course, be adding credit to the pugixml author(s) in the next README
update.