Make the windowing system chained
Have the windows clean up after themselves and restore themselves when being
chained amongst themselves.

However, DSW and Text Input aren't playing too nicely with it. DSW doesn't
exactly clean up after itself correctly, leaving it's title text behind. Text
Input has the problem of not restoring it's inner window. I have a theory that
inner windows must be backed up and restored individually. In which case, I
might have to make the window class a virtual class.

Conversely, I could just not worry about the menuing system and go on with
working on the core of the actual game. But I'm kind of a perfectionist, so
that isn't really even an option, is it?